requirements.txt
transformers 
datasets 
torchaudio 
accelerate 
librosa  
IPython 
numpy 
pydub 
youtube_dl
numpy 
torch
flask 
datasets 
transformers~=4.28.0 
tqdm
google-api-python-client
flask_cors
moviepy
python-dotenv